db/requirements.txt
psycopg2-binary>=2.9.0
rapidfuzz>=3.0.0
